+++ .juggrc.doc.ts
@@ -2,6 +2,7 @@ import { extendConfig } from '@axew/jugg';
 
 export default extendConfig({
   plugins: [
+    '@axew/jugg-plugin-vue',
     [
       '@axew/jugg-plugin-doc',
       {