+++ packages/jugg-plugin-doc/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.0.1-alpha.0
 
 - [ ] refactor: 优化配置
+- refactor: run React demo
+- refactor: run Vue demo
 - refactor: 优化路由
 - feat: 网站成功运行